refactor: :truck: merge cluster routes
