+++ apps/server/src/resources/cluster/router.ts
@@ -1,7 +1,17 @@
 import { clusterContract } from '@cpn-console/shared'
 import { addReqLogs } from '@/utils/logger.js'
 import { getAllUserClusters } from './business.js'
+import '@/types/index.js'
 import { serverInstance } from '@/app.js'
+import {
+  checkClusterProjectIds,
+  createCluster,
+  deleteCluster,
+  getClusterAssociatedEnvironments,
+  getClusterDetails as getClusterDetailsBusiness,
+  updateCluster,
+} from './admin/business.js'
+import { checkIsAdmin } from '@/utils/controller.js'
 
 export const clusterRouter = () => serverInstance.router(clusterContract, {
   listClusters: async ({ request: req }) => {
@@ -12,6 +22,71 @@ export const clusterRouter = () => serverInstance.router(clusterContract, {
     return {
       status: 200,
       body: cleanedClusters,
+    }
+  },
+
+  getClusterDetails: async ({ params, request: req }) => {
+    checkIsAdmin(req.session.user)
+    const clusterId = params.clusterId
+    const cluster = await getClusterDetailsBusiness(clusterId)
+
+    return {
+      status: 200,
+      body: cluster,
+    }
+  },
+
+  createCluster: async ({ request: req, body: data }) => {
+    checkIsAdmin(req.session.user)
+    const userId = req.session.user.id
+
+    data.projectIds = checkClusterProjectIds(data)
+    const cluster = await createCluster(data, userId, req.id)
+
+    addReqLogs({ req, message: 'Cluster créé avec succès', infos: { clusterId: cluster.id } })
+    return {
+      status: 201,
+      body: cluster,
+    }
+  },
+
+  getClusterEnvironments: async ({ request: req, params }) => {
+    checkIsAdmin(req.session.user)
+    const clusterId = params.clusterId
+    const environments = await getClusterAssociatedEnvironments(clusterId)
+
+    addReqLogs({ req, message: 'Environnements associés au cluster récupérés', infos: { clusterId } })
+    return {
+      status: 200,
+      body: environments,
+    }
+  },
+
+  updateCluster: async ({ request: req, params, body: data }) => {
+    checkIsAdmin(req.session.user)
+    const userId = req.session.user.id
+    const clusterId = params.clusterId
+
+    const cluster = await updateCluster(data, clusterId, userId, req.id)
+
+    addReqLogs({ req, message: 'Cluster mis à jour avec succès', infos: { clusterId: cluster.id } })
+    return {
+      status: 200,
+      body: cluster,
+    }
+  },
+
+  deleteCluster: async ({ request: req, params }) => {
+    checkIsAdmin(req.session.user)
+    const userId = req.session.user.id
+    const clusterId = params.clusterId
+
+    await deleteCluster(clusterId, userId, req.id)
+
+    addReqLogs({ req, message: 'Cluster supprimé avec succès', infos: { clusterId } })
+    return {
+      status: 204,
+      body: null,
     }
   },
 })
