+++ apps/server/src/resources/index-admin.ts
@@ -1,5 +1,4 @@
 import { type FastifyInstance } from 'fastify'
-import { clusterAdminRouter } from './cluster/admin/router.js'
 import { logAdminRouter } from './log/admin/router.js'
 import { projectAdminRouter } from './project/admin/router.js'
 import { quotaAdminRouter } from './quota/admin/router.js'
@@ -11,7 +10,6 @@ import { checkAdminGroup } from '@/utils/controller.js'
 
 export const apiRouterAdmin = () => async (app: FastifyInstance) => {
   app.addHook('preHandler', checkAdminGroup)
-  await app.register(serverInstance.plugin(clusterAdminRouter()), { responseValidation: true })
   await app.register(serverInstance.plugin(logAdminRouter()), { responseValidation: true })
   await app.register(serverInstance.plugin(projectAdminRouter()))
   await app.register(serverInstance.plugin(quotaAdminRouter()))
